correct paths with Windows separator
Some paths have been saved in the cache with windows separators.
These paths raise file-not-found errors in Linux.
An adjust has been made to handle this situation.
A deeper change must be made in order to handle this problem properly in a way that brasa runs in any platform without problems.